Ora loadUpdate sceglie automaticamente se usare git o composer per aggiornarsi.

+++ src/neneone/getUpdates/Wrappers/loadUpdate.php
@@ -22,6 +22,10 @@ trait loadUpdate
 {
     public function loadUpdate()
     {
-        return shell_exec('git checkout master && git fetch --all && git reset --hard origin/master && git pull && cd ../.. &&  composer dump-autoload && composer update');
+        if(file_exists(__DIR__ . '/../../../../.git')) {
+          return shell_exec('git checkout master && git fetch --all && git reset --hard origin/master && git pull && cd ../.. &&  composer dump-autoload && composer update');
+        } else {
+          return passthru('composer update');
+        }
     }
 }
